Sends a request to kill istio sidecar after the job completes
Following the new (and more strict configuration of PlatIAgro),
the job runs aside istio sidecar (to check for permissions to
access MySQL and MinIO).
However the sidecar runs forever, and the job wouldn't complete
unless a sign is sent to kill the sidecar.
This commits adds the request that terminates the sidecar:
POST http://localhost:15020/quitquitquit